+++ docs/changelog.rst
@@ -3,7 +3,8 @@ Changelog
 #########
 
 0.4.1
-    + Fix force_text to try force_unicode first – in feeds.
+    + Add tags to cms toolbar menu
+    + Fix force_text to try force_unicode first – in feeds
     + Separate managers from models
     + Fix post detail view returning multiple objects when different language
       posts have the same slug